backend/pyproject.toml
[project]
name = "api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "celery>=5.6.2",
    "fastapi>=0.135.1",
    "numpy>=2.4.2",
    "pillow>=12.1.1",
    "psycopg2-binary>=2.9.11",
    "pydantic-settings>=2.13.1",
    "python-multipart>=0.0.22",
    "redis>=7.3.0",
    "sqlalchemy>=2.0.48",
    "torch>=2.10.0",
    "torchvision>=0.25.0",
    "ultralytics>=8.4.21",
    "uvicorn[standard]>=0.41.0",
]
